#14424
feat(auto-pair): make autopair work with custom strings instead of chars

Make AutoPair a prefix-tree map data structure.
Adapt all the clients.
Add multiple-char autopairs for python and latex as proper examples.

fix(clippy): apply clippy's fixes

fix(clippy): resolve runaway lint in helix-term

fix(auto-pairs): fix the matching of the closing-side of the pair from the cursor on. As a consequence, this fixes the indentation of a newline inserted between the '\(' and '\)' latex pair.

fix: construct AutoPairs correctly independently of the order of the given pairs.

test: add first test for auto-pairs with multiple characters. The last char of the opening pair triggers the completion.

test: add test append. Divide insert tests for multichar pairs and singlechar pairs